Moves modules into introductions module.
This commit moves the `intro` and `voice` modules into the
`introductions` module with `intro` renamed to `queries`. This seems
sensible in terms of structure but I'm still unsure of how to best move
the rest of the introductions business logic out of the `app` module.
