+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod app;
 pub mod database_setup;
-pub mod intro;
-pub mod voice;
+pub mod introductions;
